Move stuff to libraries. Fun fact: self.saveit was previously never modified, so this never worked
