+++ bin/gftools-fix-glyph-private-encoding.py
@@ -17,9 +17,11 @@
 # See AUTHORS.txt for the list of Authors and LICENSE.txt for the License.
 #
 import argparse
-import copy
 import os
 from fontTools import ttLib
+from gftools.fix import fix_pua
+from gftools.utils import get_unencoded_glyphs
+
 
 description = 'Fixes TTF unencoded glyphs to have Private Use Area encodings'
 
@@ -29,25 +31,6 @@ parser.add_argument('ttf_font', nargs='+',
 parser.add_argument('--autofix', action="store_true",
                     help='Apply autofix. '
                          'Otherwise just check if there are unencoded glyphs')
-
-
-def get_unencoded_glyphs(font):
-    """ Check if font has unencoded glyphs """
-    cmap = font['cmap']
-
-    new_cmap = cmap.getcmap(3, 10)
-    if not new_cmap:
-        for ucs2cmapid in ((3, 1), (0, 3), (3, 0)):
-            new_cmap = cmap.getcmap(ucs2cmapid[0], ucs2cmapid[1])
-            if new_cmap:
-                break
-
-    if not new_cmap:
-        return []
-
-    diff = list(set(font.getGlyphOrder()) -
-                set(new_cmap.cmap.values()) - {'.notdef'})
-    return [g for g in diff[:] if g != '.notdef']
 
 
 class AddSPUAByGlyphIDToCmap(object):
@@ -62,38 +45,7 @@ class AddSPUAByGlyphIDToCmap(object):
             self.font.save(self.path + ".fix")
 
     def fix(self):
-        unencoded_glyphs = get_unencoded_glyphs(self.font)
-        if not unencoded_glyphs:
-            return
-
-        ucs2cmap = None
-        cmap = self.font["cmap"]
-
-        # Check if an UCS-2 cmap exists
-        for ucs2cmapid in ((3, 1), (0, 3), (3, 0)):
-            ucs2cmap = cmap.getcmap(ucs2cmapid[0], ucs2cmapid[1])
-            if ucs2cmap:
-                break
-        # Create UCS-4 cmap and copy the contents of UCS-2 cmap
-        # unless UCS 4 cmap already exists
-        ucs4cmap = cmap.getcmap(3, 10)
-        if not ucs4cmap:
-            cmapModule = ttLib.getTableModule('cmap')
-            ucs4cmap = cmapModule.cmap_format_12(12)
-            ucs4cmap.platformID = 3
-            ucs4cmap.platEncID = 10
-            ucs4cmap.language = 0
-            if ucs2cmap:
-                ucs4cmap.cmap = copy.deepcopy(ucs2cmap.cmap)
-            cmap.tables.append(ucs4cmap)
-        # Map all glyphs to UCS-4 cmap Supplementary PUA-A codepoints
-        # by 0xF0000 + glyphID
-        ucs4cmap = cmap.getcmap(3, 10)
-        for glyphID, glyph in enumerate(self.font.getGlyphOrder()):
-            if glyph in unencoded_glyphs:
-                ucs4cmap.cmap[0xF0000 + glyphID] = glyph
-        self.font['cmap'] = cmap
-        return True
+        self.saveit = fix_pua(self.font)
 
 
 def main():
